refactor(scan): extract garak async bridge into _bridge.py
