+++ libs/giskard-scan/src/giskard/scan/integrations/garak/_generator.py
@@ -1,37 +1,12 @@
 import asyncio
 import uuid
-from collections.abc import Coroutine
-from typing import Any, override
+from typing import override
 
 from garak.attempt import Conversation, Message
 from garak.generators.base import Generator
 from giskard.checks import DatasetInputGenerator, Interact, Target, Trace
 
-
-def _await_on_loop[T](
-    coro: Coroutine[Any, Any, T],
-    loop: asyncio.AbstractEventLoop | None,
-) -> T:
-    """Run *coro* without spawning a nested event loop in a worker thread.
-
-    Garak calls ``_call_model`` synchronously, often from ``asyncio.to_thread``
-    workers. ``asyncio.run`` there creates ephemeral loops that break
-    ``asyncio.Lock`` in ``DatasetInputGenerator`` when probes run in parallel.
-    When *loop* is the scan's running loop, schedule the coroutine there via
-    ``run_coroutine_threadsafe`` instead.
-    """
-    if loop is None:
-        return asyncio.run(coro)
-    future = asyncio.run_coroutine_threadsafe(coro, loop)
-    return future.result()
-
-
-def _conv_uuid(conversation: Conversation) -> str | None:
-    for turn in conversation.turns:
-        notes = turn.content.notes
-        if notes and notes.get("uuid"):
-            return notes["uuid"]
-    return None
+from ._bridge import _await_on_loop, _conv_uuid
 
 
 class TargetGenerator[InputType, OutputType, TraceType: Trace](  # pyright: ignore[reportMissingTypeArgument]
